Move TMR AXI M test
Internal-tag: [#101876]
